requirements.txt
fastapi==0.95.2
uvicorn==0.23.2
pymongo==4.16.0
python-dotenv==1.2.1
pydantic[email]==2.12.5